+++ src/main/java/store/model/DateRange.java
@@ -1,4 +1,4 @@
-package store.domain;
+package store.model;
 
 import camp.nextstep.edu.missionutils.DateTimes;
 import java.time.LocalDate;
