+++ src/test/java/store/parser/ProductStatusParserTest.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import store.domain.ProductStatus;
+import store.model.ProductStatus;
 import store.utility.CSVReader;
 
 class ProductStatusParserTest {
