+++ src/test/java/store/parser/PromotionParserTest.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import store.domain.Promotion;
+import store.model.Promotion;
 import store.utility.CSVReader;
 
 class PromotionParserTest {
